- Changed PrimaryKey values from Int to Long - Made WorkoutDb insert function return id type Long for both having the new record id and its successful insertion - getWorkoutDbAndRoundsDb returns LiveData - Reimplemented ViewModel in SavedFragment, refactored recycler adapter code, - Updated SavedFragmentViewModel - Fixed Adapter
* SavedAdapter.kt was only modified not to throw an error - don't know its purpose or anything

+++ app/src/main/java/com/appdavide/roundtimer/data/WorkoutDb/WorkoutDb.kt
@@ -9,6 +9,6 @@ class WorkoutDb(
 
     @ColumnInfo(name = "name") var name : String,
 
-    @PrimaryKey(autoGenerate = true) val id: Int = 0
+    @PrimaryKey(autoGenerate = true) val id: Long = 0
 
     )
